requirements.txt
fastapi==0.136.0
httpx==0.28.1
uvicorn==0.44.0
spacy==3.8.14
sentence-transformers==5.4.1
nltk==3.9.4
rouge-score==0.0.4
bert-score==0.3.13
datasets==4.8.4
numpy==2.4.4
scipy==1.17.1
matplotlib==3.10.8
PyYAML==6.0.3
scikit-learn==1.8.0
pytest==9.0.3
llama-cpp-python==0.3.20; python_version < "3.13"
